ml/requirements_phase8.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python-headless>=4.8.0.74
mediapipe>=0.10.9
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
